refactor 페이지네이션 이름 수정
